requirements.txt
aiogram>=3.6.0
python-dotenv>=1.0.1
